+++ src/anomalib/models/image/__init__.py
@@ -23,7 +23,8 @@ Example:
     >>> predictions = engine.predict(model=model, datamodule=datamodule)  # doctest: +SKIP
 
 Available Models:
-    - :class: `AnomalyDINO`: Boost Memorybank Models with DINOv2
+    - :class:`AnomalyVFM`: Transforming Vision Foundation Models into Zero-Shot Anomaly Detectors
+    - :class:`AnomalyDINO`: Boost Memorybank Models with DINOv2
     - :class:`Cfa`: Contrastive Feature Aggregation
     - :class:`Cflow`: Conditional Normalizing Flow
     - :class:`CFM`: Crossmodal Feature Mapping
@@ -52,6 +53,7 @@ Available Models:
 """
 
 from .anomaly_dino import AnomalyDINO
+from .anomalyvfm import AnomalyVFM
 from .cfa import Cfa
 from .cflow import Cflow
 from .cfm import CFM
@@ -79,6 +81,7 @@ from .vlm_ad import VlmAd
 from .winclip import WinClip
 
 __all__ = [
+    "AnomalyVFM",
     "AnomalyDINO",
     "Cfa",
     "Cflow",